orders/src/routes/delete.ts
import express, { Request, Response } from 'express';
import { requireAuth, NotFoundError, NotAuthorizedError } from '@catevikatickets/common';
import { Order, OrderStatus } from '../models/order';
import { OrderCancelledPublisher } from '../events/publishers/order-cancelled-publisher';
import { natsWrapper } from '../nats-wrapper';

const router = express.Router();

// It should have been better to use a PATCH instead of DELETE and change the status to 200
router.delete('/api/orders/:orderId', requireAuth, async (req: Request, res: Response) => {
  const { orderId } = req.params;

  const order = await Order.findById(orderId).populate('ticket');

  if (!order) {
    throw new NotFoundError();
  }

  if (order.userId !== req.currentUser!.id) {
    throw new NotAuthorizedError();
  }

  order.status = OrderStatus.Cancelled;
  await order.save();

  // publishing an event saying that the order was cancelled
  new OrderCancelledPublisher(natsWrapper.client).publish({
    id: order.id,
    version: order.version,
    ticket: {
      id: order.ticket.id
    }
  })


  // It should have been better to use a PATCH instead of DELETE and change the status to 200
  res.status(204).send(order);
});

export { router as deleteOrderRouter };